Add support for Updating Maven Pom Dependencies
The command will run trough current pom and all children to check if there are
newer version of the defined artifacts available in the Maven repositories.

If one is found, the user will be prompted to see if he wants to update,
ignore or see a complete list of available newer versions.

Two events are fired by this Command:

* UpdatedDependency

  Fired when the user has choosen to update a given dependency.
  The From and Two Artifact Info is given in the Event. This can be used to
  e.g. provide Release Notes for a given Artifact.

* UpdatingDependency

  Fired when the user has choosen to update a given dependency, but before
  the change has been written to disk. The event allow for Observers to
  veto the update. This can be used by other Plugins to warn / abort
  updates that would cause an invalid configuration.

Usage$ maven update