FORGE-813 Generate FETCH JOINs in the REST resource JPQL queries.
The JOIN FETCH expressions ensure that  immediate collections are fetched. This fits in well, with the upcoming changes for FORGE-606, where the first immediate properties of an object are serialized, thus avoiding the open-session in view.

Collections at a deeper level in the related entities may be fetched, depending on the EAGER/LAZY fetchType values on the nested collections.

Also replaced Seam Render usage with FreeMarker.