FORGE-870 Support to allow non-generated Ids to be visible.
The ForgeInspector is enhanced to provide inspection results indicating Ids that are generated values. The GeneratedValueInspectionResultProcessor uses the information to hide or unhide the fields. This overrides the JPAInspector's inspection result on whether @Ids should be hidden or not. All Metawidgets used by the faces scaffold now use this inspection result processor.