FORGE-873 Overrides behavior of default HtmlWidgetBuilder for rendering h:selectOneMenu widgets.
The selectOneMenu widgets are generated with labels whenever these happen to be lookups on other entities. The behavior of depending on toString() is done away with, since this is not dependable. Additionally, due to Metawidget limitations, it is not possible to determine whether a toString() method is defined on an entity, and then adjust the generated widget. Hence, we now generate labels containing the primary keys of the other side of the relationship.

This is also done for the search results when a reference to another entity is displayed. This is also done when a link is displayed to another entity in the view screen.