FORGE-873 Overrides behavior of default HtmlWidgetBuilder
This renders h:selectOneMenu widgets differently.

The selectOneMenu widgets are generated with EL expressions that invoke a tag library to determine how labels should be displayed. The tag library determines if toString() is defined by the class of the instance and invokes it based on the method's presence. Additionally, due to Metawidget limitations, it is not possible to determine whether a toString() method is defined on an entity, and then adjust the generated widget. Hence, specialized EL expressions involving the tag library are generated instead of relying on Metawidget to generate the right expressions.

This is also done for the search results when a reference to another entity is displayed. This is also done when a link is displayed to another entity in the view screen.