FORGE-93 Return correct qualified path on Windows
The fix was to ensure that the absolute path is returned for File resources.

Also corrected the manner in which the prompt displayed, since the entire path was displayed on Windows due to the above described fix.

Also incidentally fixed an issue related to FORGE-554 involving relative paths. The earlier fix only resolved the issue involving absolute paths.