FORGE-1100 Use the in-flight project instance for configuration
The in-flight project instance that is being setup will now be
used whenever Facets are being registered. The underlying
configuration file of this project instance will be used when
requested.

This will ensure that @Inject 'ed Configuration instances can
be used inside Facet.install() and Facet.isInstalled() methods.
These injected Configuration instances will no longer use the
User scoped configuration store if there is a project
associated with the facet.