FORGE-1098 Detect the Facet installation phase for scaffolds.
The InstallFacet event for the scaffold is fired when the
setup command is executed. This is necessary otherwise the
scaffold plugin would not detect the installed ScaffoldProvider
aka the facet of the scaffold provider.

Additionally, the currently installed scaffold for the project is
used for scaffolding only when other commands are executed and no
other scaffold provider is detected.