FORGE-1134 Fixing resolution failures for secured repositories
If a repository is defined along with a server configuration
element in settings.xml, then any username and password settings
configured, are used for that repository. Additionally, any
activeByDefault profiles are treated as active profiles with
this fix.

Passwords for mirrored repositories are determined at runtime
after Maven+Aether has performed a merge of the Maven model to
determine whether a mirror should used. The password for the
mirror is then retrieved at runtime if a mirror is involved.

Also fixed tests for ShellMessages failures. Delayed the
invocation of System.gc until all temp directories are deleted,
to improve build performance.