Replaced ShrinkWrap Descriptors class with Forge class.
The default classloader used to load the descriptors in ShrinkWrap Descriptors
is the current context classloader. This does not work in Forge 2, given that
classes are loaded from different classloaders. The current implementation was
throwing IllegalStateException, because the Descriptor resource was being loaded
from the "facets" addon, not the javaee addon.