FORGE-1261: Added Log4j as a dependency in the freemarker modules.xml
The Hibernate tools plugin already bundles freemarker as a new dependency. Freemarker has a built-in logging facade
system that delegates to the underlying logging implementations (much like commons-logging). When the Hibernate tools
plugin is installed, log4j (which is bundled with this plugin as well) kicks in and freemarker chooses the Log4j
implementation. This is not handled correctly in the module.xml, given that declares that it depends on nothing but the javax.api.