JBIDE-18330 Fix slowness in script execution on Eclipse and Win.
A new ForgeTerminal is created and supplied for running the lines
in the script. This ensures that a new WindowsTerminal instance
is not created. Using a WindowsTerminal prevents ANSI escape codes
from being sent, preventing styles from being correctly applied.

As a result, on Windows, only one style range is used prior to this
commit, ensuring that the Document widget update take O(n), where
n is the number of charatcers in it.